refactor: outsourced bussines logic to controller from routes
