+++ controllers/products.js
@@ -1,0 +1,27 @@
+const products = [];
+
+exports.getAddProducts = (req, res, next) => {
+  res.render('add-product', {
+    pageTitle: 'Add Product',
+    path: '/admin/add-product',
+    formsCSS: true,
+    productCSS: true,
+    activeAddProduct: true
+  });
+};
+
+exports.getPostProducts = (req, res, next) => {
+  products.push({ title: req.body.title });
+  res.redirect('/');
+};
+
+exports.getProducts = (req, res, next) => {
+  res.render('shop', {
+    prods: products,
+    pageTitle: 'Shop',
+    path: '/',
+    hasProducts: products.length > 0,
+    activeShop: true,
+    productCSS: true
+  });
+};
