+++ index.js
@@ -10,8 +10,9 @@ app.set('view engine', 'ejs');
 /* folder to be used for views */
 app.set('views', 'views');
 
-const adminData = require('./routes/admin');
+const adminRoutes = require('./routes/admin');
 const shopRoutes = require('./routes/shop');
+const { pageNotFound } = require('./controllers/error404');
 
 /* to parse body of the request and to set it inside req.body */
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -19,14 +20,11 @@ app.use(bodyParser.urlencoded({ extended: false }));
 /* to serve static files like css or png's - by default will not accessible */
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/admin', adminData.routes);
+app.use('/admin', adminRoutes.routes);
 app.use(shopRoutes);
 
-app.use((req, res, next) => {
-  res.status(404).render('404', { pageTitle: 'Page Not Found' });
-});
-
+app.use(pageNotFound);
 
 app.listen(3000, () => {
-  console.log('App started successfully')
+  console.log('App started successfully');
 });
