+++ routes/admin.js
@@ -1,29 +1,13 @@
-const path = require('path');
-
 const express = require('express');
 
-const rootDir = require('../util/path');
+const { getAddProducts, getPostProducts } = require('../controllers/products');
 
 const router = express.Router();
 
-const products = [];
-
 // /admin/add-product => GET
-router.get('/add-product', (req, res, next) => {
-  res.render('add-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
-    formsCSS: true,
-    productCSS: true,
-    activeAddProduct: true
-  });
-});
+router.get('/add-product', getAddProducts);
 
 // /admin/add-product => POST
-router.post('/add-product', (req, res, next) => {
-  products.push({ title: req.body.title });
-  res.redirect('/');
-});
+router.post('/add-product', getPostProducts);
 
 exports.routes = router;
-exports.products = products;
