+++ routes/shop.js
@@ -1,22 +1,9 @@
-const path = require('path');
-
 const express = require('express');
 
-const rootDir = require('../util/path');
-const adminData = require('./admin');
+const { getProducts } = require('../controllers/products');
 
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-  const products = adminData.products;
-  res.render('shop', {
-    prods: products,
-    pageTitle: 'Shop',
-    path: '/',
-    hasProducts: products.length > 0,
-    activeShop: true,
-    productCSS: true
-  });
-});
+router.get('/', getProducts);
 
 module.exports = router;
